Move hostrouter code to its own file
