+++ pkg/middleware/request_host.go
@@ -1,0 +1,75 @@
+/*
+Copyright (c) 2016-Present https://github.com/go-chi authors
+
+MIT License
+
+Permission is hereby granted, free of charge, to any person obtaining a copy of
+this software and associated documentation files (the "Software"), to deal in
+the Software without restriction, including without limitation the rights to
+use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+the Software, and to permit persons to whom the Software is furnished to do so,
+subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*/
+
+// requestHost and parseForwarded are taken from https://github.com/go-chi/hostrouter
+// (permanent link: https://github.com/go-chi/hostrouter/blob/7bff2694dfd99a31a89c62e5f8a2d9ec2d71da8e/hostrouter.go)
+
+// Package middleware //
+package middleware
+
+import (
+	"net/http"
+	"strings"
+)
+
+func requestHost(r *http.Request) (host string) {
+	// not standard, but most popular
+	host = r.Header.Get("X-Forwarded-Host")
+	if host != "" {
+		return
+	}
+
+	// RFC 7239
+	host = r.Header.Get("Forwarded")
+	_, _, host = parseForwarded(host)
+	if host != "" {
+		return
+	}
+
+	// if all else fails fall back to request host
+	host = r.Host
+	return
+}
+
+func parseForwarded(forwarded string) (addr, proto, host string) {
+	if forwarded == "" {
+		return
+	}
+	for _, forwardedPair := range strings.Split(forwarded, ";") {
+		if tv := strings.SplitN(forwardedPair, "=", 2); len(tv) == 2 {
+			token, value := tv[0], tv[1]
+			token = strings.TrimSpace(token)
+			value = strings.TrimSpace(strings.Trim(value, `"`))
+			switch strings.ToLower(token) {
+			case "for":
+				addr = value
+			case "proto":
+				proto = value
+			case "host":
+				host = value
+			}
+
+		}
+	}
+	return
+}
